agents/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai
langchain-tavily
edgartools
finnhub-python
yfinance
openai
pymongo
psycopg2-binary==2.9.9
fastapi==0.111.0